# swagger.yml
openapi: 3.0.0
info:
  title: "Trails Management API"
  description: "API for managing trails, features, location points, and user authentication. This API allows users to log in, manage trails, add features, update location points, and more. Administrators and standard users have different access levels."
  version: "1.1.2"
servers:
  - url: "http://127.0.0.1:8000/api"
paths:
  /login:
    post:
      summary: "Authenticate user and generate JWT token"
      description: "This endpoint allows users to log in by providing a username (email) and password. Upon successful authentication, a JWT token is returned, along with the user's role."
      operationId: "routes.login"
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  description: "User's email address"
                password:
                  type: string
                  description: "User's password"
      responses:
        "200":
          description: "Login successful"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Login successful"
                  token:
                    type: string
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  role:
                    type: string
                    example: "Administrator"
        "400":
          description: "Bad Request - Missing username or password"
        "401":
          description: "Unauthorized - Invalid credentials"
        "500":
          description: "Authentication service unavailable"
  /trails/admin:
    get:
      summary: "Retrieve all trails (Administrator view)"
      description: "This endpoint allows administrators to retrieve a list of all trails."
      operationId: "routes.admin_trails"
      tags:
        - Trails
      security:
        - BearerAuth: []
      responses:
        "200":
          description: "List of trails for administrators"
          content:
            application/json:
              schema:
                type: object
                properties:
                  trails:
                    type: array
                    items:
                      $ref: "#/components/schemas/Trail"
        "401":
          description: "Unauthorized - Missing or invalid token"
        "403":
          description: "Forbidden - User does not have administrative access"
  /trails/standard:
    get:
      summary: "Retrieve all trails (Standard view)"
      description: "This endpoint allows standard users to retrieve a list of trails."
      operationId: "routes.standard_trails"
      tags:
        - Trails
      security:
        - BearerAuth: []
      responses:
        "200":
          description: "List of trails for standard users"
          content:
            application/json:
              schema:
                type: object
                properties:
                  trails:
                    type: array
                    items:
                      $ref: "#/components/schemas/Trail"
        "401":
          description: "Unauthorized - Missing or invalid token"
        "403":
          description: "Forbidden - User does not have access"
  /trails/admin/{trail_id}:
    get:
      summary: "Retrieve a single trail by ID (Administrator view)"
      description: "Allows administrators to retrieve details of a specific trail using its ID."
      operationId: "routes.trail_by_id_admin"
      tags:
        - Trails
      parameters:
        - name: "trail_id"
          in: path
          description: "ID of the trail"
          required: true
          schema:
            type: integer
      security:
        - BearerAuth: []
      responses:
        "200":
          description: "Trail details for administrators"
          content:
            application/json:
              schema:
                type: object
                properties:
                  trail:
                    $ref: "#/components/schemas/Trail"
        "401":
          description: "Unauthorized - Missing or invalid token"
        "403":
          description: "Forbidden - User does not have administrative access"
        "404":
          description: "Trail not found"
  /trails/standard/{trail_id}:
    get:
      summary: "Retrieve a single trail by ID (Standard view)"
      description: "Allows standard users to retrieve details of a specific trail using its ID."
      operationId: "routes.trail_by_id_standard"
      tags:
        - Trails
      parameters:
        - name: "trail_id"
          in: path
          description: "ID of the trail"
          required: true
          schema:
            type: integer
      security:
        - BearerAuth: []
      responses:
        "200":
          description: "Trail details for standard users"
          content:
            application/json:
              schema:
                type: object
                properties:
                  trail:
                    $ref: "#/components/schemas/Trail"
        "401":
          description: "Unauthorized - Missing or invalid token"
        "403":
          description: "Forbidden - User does not have access"
        "404":
          description: "Trail not found"
  /trails/add:
    post:
      summary: "Add a new trail"
      description: "Allows administrators to add a new trail to the system."
      operationId: "routes.add_trail"
      tags:
        - Trails
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrailCreate"
      security:
        - BearerAuth: []
      responses:
        "200":
          description: "Trail added successfully"
        "400":
          description: "Bad Request - Invalid trail data"
        "401":
          description: "Unauthorized - Missing or invalid token"
  /trails/update/{trail_id}:
    put:
      summary: "Update a trail"
      description: "Allows administrators to update the details of an existing trail."
      operationId: "routes.edit_trail"
      tags:
        - Trails
      parameters:
        - name: "trail_id"
          in: path
          description: "ID of the trail"
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrailCreate"
      security:
        - BearerAuth: []
      responses:
        "200":
          description: "Trail updated successfully"
        "400":
          description: "Bad Request - Invalid trail data"
        "401":
          description: "Unauthorized - Missing or invalid token"
        "404":
          description: "Trail not found"
  /trails/delete/{trail_id}:
    delete:
      summary: "Delete a trail"
      description: "Allows administrators to delete a trail using its ID."
      operationId: "routes.remove_trail"
      tags:
        - Trails
      parameters:
        - name: "trail_id"
          in: path
          description: "ID of the trail"
          required: true
          schema:
            type: integer
      security:
        - BearerAuth: []
      responses:
        "200":
          description: "Trail deleted successfully"
        "401":
          description: "Unauthorized - Missing or invalid token"
        "404":
          description: "Trail not found"
  /trails/{trail_id}/features:
    post:
      summary: "Add features to a trail"
      description: "Allows administrators to add one or more features to a trail."
      operationId: "routes.add_trail_features"
      tags:
        - Features
      parameters:
        - name: "trail_id"
          in: path
          description: "ID of the trail"
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
      security:
        - BearerAuth: []
      responses:
        "200":
          description: "Features added successfully"
        "400":
          description: "Bad Request - Invalid feature data"
        "401":
          description: "Unauthorized - Missing or invalid token"
        "404":
          description: "Trail not found"
  /trails/{trail_id}/features/{feature_id}:
    delete:
      summary: "Remove a feature from a trail"
      description: "Allows administrators to remove a specific feature from a trail."
      operationId: "routes.remove_trail_feature"
      tags:
        - Features
      parameters:
        - name: "trail_id"
          in: path
          description: "ID of the trail"
          required: true
          schema:
            type: integer
        - name: "feature_id"
          in: path
          description: "ID of the feature"
          required: true
          schema:
            type: integer
      security:
        - BearerAuth: []
      responses:
        "200":
          description: "Feature removed successfully"
        "401":
          description: "Unauthorized - Missing or invalid token"
        "404":
          description: "Trail or feature not found"
  /trails/{trail_id}/location-points:
    put:
      summary: "Update location points for a trail"
      description: "Allows administrators to update the location points of a trail."
      operationId: "routes.edit_location_points"
      tags:
        - Location Points
      parameters:
        - name: "trail_id"
          in: path
          description: "ID of the trail"
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/LocationPoint"
      security:
        - BearerAuth: []
      responses:
        "200":
          description: "Location points updated successfully"
        "400":
          description: "Bad Request - Invalid location point data"
        "401":
          description: "Unauthorized - Missing or invalid token"
        "404":
          description: "Trail not found"
  /trails/{trail_id}/location-points/{location_point_id}:
    delete:
      summary: "Remove a location point from a trail"
      description: "Allows administrators to remove a specific location point from a trail."
      operationId: "routes.delete_location_point"
      tags:
        - Location Points
      parameters:
        - name: "trail_id"
          in: path
          description: "ID of the trail"
          required: true
          schema:
            type: integer
        - name: "location_point_id"
          in: path
          description: "ID of the location point"
          required: true
          schema:
            type: integer
      security:
        - BearerAuth: []
      responses:
        "200":
          description: "Location point removed successfully"
        "401":
          description: "Unauthorized - Missing or invalid token"
        "404":
          description: "Trail or location point not found"
components:
  schemas:
    Account:
      type: object
      properties:
        AccountID:
          type: integer
        Email_address:
          type: string
          description: "Email address of the account"
        Role:
          type: string
          description: "Role of the account user (Administrator or Standard)"
    Trail:
      type: object
      properties:
        TrailID:
          type: integer
        Trail_name:
          type: string
        Trail_Summary:
          type: string
        Trail_Description:
          type: string
        Difficulty:
          type: string
        Location:
          type: string
        Length:
          type: number
        Elevation_gain:
          type: integer
        Route_type:
          type: string
        OwnerID:
          type: integer
    TrailCreate:
      type: object
      required:
        - Trail_name
        - Location
        - Length
        - Difficulty
        - Route_type
      properties:
        Trail_name:
          type: string
        Trail_Summary:
          type: string
        Trail_Description:
          type: string
        Difficulty:
          type: string
        Location:
          type: string
        Length:
          type: number
        Elevation_gain:
          type: integer
        Route_type:
          type: string
    Feature:
      type: object
      properties:
        FeatureID:
          type: integer
        Feature_name:
          type: string
          description: "Name of the feature"
    LocationPoint:
      type: object
      properties:
        LocationPointID:
          type: integer
        Latitude:
          type: number
        Longitude:
          type: number
        Description:
          type: string
    TrailFeature:
      type: object
      properties:
        TrailID:
          type: integer
        FeatureID:
          type: integer
    TrailLocationPoint:
      type: object
      properties:
        TrailID:
          type: integer
        LocationPointID:
          type: integer
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      x-bearerInfoFunc: config.bearer_info_func